fix(frontend): resolve clippy warnings with -D warnings
Clean up frontend lints for WASM builds, disable native test harness on the bin target, and remove unused API helper code.

+++ frontend/src/components/song_viewer.rs
@@ -22,7 +22,7 @@ pub fn song_viewer(props: &Props) -> Html {
         props.override_key.as_ref(),
         props.override_representation.as_ref(),
         None,
-        Some(scale as f32),
+        Some(scale),
     );
     html! {
         <div
